requirements.txt
asgiref==3.7.2
Django==4.2.10
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
